minor bugfixes and rename from refl_bands.py
